requirements.txt
appdirs==1.4.4
astropy==6.1.0
brotlipy==0.7.0
certifi==2023.5.7
colorama==0.4.6
dill==0.3.6
fonttools==4.25.0
ipython==8.13.2
Jinja2==3.1.2
joblib==1.2.0
jplephem==2.18
MarkupSafe==2.1.3
matplotlib==3.7.1
mpl-tools==0.2.57
munkres==1.1.4
multiprocessing-on-dill==3.5.0a4
networkx==3.1
numpy==1.26.4
pandas==1.5.3
patlib==0.3.5
Pillow==9.4.0
ply==3.11
pyorb==0.5.4
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pyyaml==6.0.1
scikit-learn==1.2.2
scipy==1.13.0
seaborn==0.12.2
skyfield==1.46
struct-tools==0.2.7
tabulate==0.9.0
threadpoolctl==2.2.0
tqdm==4.65.0